proxy: Move proxy port management to pkg/proxy/proxyports
Separate proxy port maintenance functions to their own package, so that
access to the private fields is clear and protected with their own mutex.

This solves lock ordering issue seen with LOCKDEBUG build between CEC
listener port allocation (upsertListener takes a lock, parseResources
calls AckProxyPort that took proxy.mutex) and proxy redirect creation,
where proxy.mutex is taken first, and AddListener takes a lock after).

With the new structure CEC parser does not need to take proxy.mutex (but
ProxyPorts.mutex), while proxy redirect creation calls AddListener
without holding the new ProxyPorts.mutex.

This also allows AddListener callback function to ack or reset a proxy
port without taking the proxy.mutex.
