+++ pkg/proxy/proxyports/netstat.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package proxy
+package proxyports
 
 import (
 	"bytes"
